% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planner.R
\name{adjust_to_weekday}
\alias{adjust_to_weekday}
\title{Adjust Dates to Weekdays}
\usage{
adjust_to_weekday(date)
}
\arguments{
\item{date}{A vector of class /code{Date} (base R).}
}
\value{
A /code{Date} vector the same length as /code{date}, adjusted so that
weekends and Fridays are shifted to the appropriate weekday.
}
\description{
Adjusts a given date so that any date falling on
a weekend or Friday is moved to the closest workday:
}
\details{
\itemize{
\item Sunday moves to Monday (+1)
\item Saturday moves to Monday (+2)
\item Friday moves to Thursday (-1)
\item All other days remain unchanged
}
}
\examples{
adjust_to_weekday(as.Date(c("2023-10-14", "2023-10-15", "2023-10-13"))) # Sat, Sun, Fri
}
