% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planner.R
\name{remake_sch}
\alias{remake_sch}
\title{Remake Schedule by Collection Groups}
\usage{
remake_sch(plan_master_curated)
}
\arguments{
\item{plan_master_curated}{A data.frame or tibble with at least 'Physical Tag' and 'Collection_Date' columns.}
}
\value{
A modified data.frame or tibble with rescheduled collection dates.
}
\description{
Organize collection dates into groups based on sample count and day difference.
}
