% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planner.R
\name{update_collection}
\alias{update_collection}
\title{Update a Database Collection by Adding New Files}
\usage{
update_collection(dir, database, id_name_alt = NA)
}
\arguments{
\item{dir}{Directory path containing Excel files.}

\item{database}{The existing database data frame serving as a template for columns.}

\item{id_name_alt}{Optional alternative column name to be renamed to 'Physical Tag'. Default is NA (no renaming).}
}
\value{
A data frame combining the original database and new rows from the directory.
}
\description{
Reads in all Excel files from a directory, optionally renaming an ID column, matches columns to a template, and combines them with an existing database.
}
\examples{
# update_collection("data", database, id_name_alt = "OtherTag")
}
